Create README.md
The "Smart Sanitization System with Face Mask Detection" is a new concept that aims at upgrading hygiene protocols through the addition of facial recognition and mask detection and automated sanitizer dispensing. It uses computer vision technology to recognize human faces and to ascertain if the person is wearing a mask.